feat(dashboard): Home selector with confirm instead of Edit dialog (LFE-10693)
Home's only control is the dashboard selector in the header; picking a
different dashboard stages the change behind a confirmation (it applies
to everyone in the project). No Edit button on Home — editing lives in
the Dashboards section (clone-first for Langfuse-managed). Home tiles
stay read-only.
